@@ -4,7 +4,7 @@ description: "A new Flutter project."
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.7.0
 
 dependencies:
   flutter:
@@ -18,7 +18,7 @@ dependencies:
 
   # for login function
   google_sign_in: ^6.2.2
-  firebase_auth: ^5.4.1
+  firebase_auth: ^5.4.2
   scrollable_positioned_list: ^0.3.8
 
   # suitable with bloc
@@ -28,20 +28,28 @@ dependencies:
   equatable: ^2.0.7
 
   # for local database
-  sqflite: 2.4.1
+  sqflite: 2.4.2
+  shared_preferences: ^2.5.2
 
   # support for create model
-  json_serializable: ^6.9.3
+  json_serializable: ^6.9.4
   json_annotation: ^4.9.0
 
-  # for view image 
+  # for view support ui 
   shimmer: ^3.0.0
   carousel_slider: ^5.0.0
+  dotted_border: ^2.1.0
+
+  # for function watch trailer from youtube
+  youtube_player_flutter: ^9.1.1
 
   # I use dio, because i like dio's api implementation
   # Retrofit It allows define API endpoints using annotations,
   dio: ^5.8.0+1
   retrofit: ^4.4.2
+
+  #format date time and other formats
+  intl: ^0.20.2
 
 dev_dependencies:
   flutter_test:
@@ -50,14 +58,19 @@ dev_dependencies:
   flutter_lints: ^5.0.0
 
   # for auto generate
-  build_runner: ^2.4.13
-  retrofit_generator: ^9.1.7
+  build_runner: ^2.4.15
+  retrofit_generator: ^9.1.9
+
+  # object builder state management
+  freezed: ^2.5.8
+  freezed_annotation: ^2.4.4
 
 flutter:
   uses-material-design: true
 
   assets:
     - assets/images/
+    - assets/images/cines/
     - assets/fonts/
     - assets/vectors/
 
@@ -73,4 +86,8 @@ flutter:
         - asset: assets/fonts/Satoshi-Medium.ttf
           weight: 500
         - asset: assets/fonts/Satoshi-Regular.ttf
+          weight: 400
+    - family: Oswald
+      fonts:
+        - asset: assets/fonts/Oswald-Regular.ttf
           weight: 400
