@@ -10,7 +10,8 @@ import 'package:cinema_booking/common/widgets/scroll_list/hoz_list_view.dart';
 import 'package:cinema_booking/common/widgets/space/widget_spacer.dart';
 import 'package:cinema_booking/core/configs/theme/app_color.dart';
 import 'package:cinema_booking/core/configs/theme/app_font.dart';
-import 'package:cinema_booking/domain/entities/movies/movies.dart';
+import 'package:cinema_booking/domain/entities/response/home.dart';
+import 'package:cinema_booking/presentation/router.dart';
 import 'package:flutter/material.dart';
 
 class WidgetHomePosters extends StatefulWidget {
@@ -59,8 +60,11 @@ class _WidgetHomePostersState extends State<WidgetHomePosters> {
           Spacer(),
           TextButton(
             onPressed: () {},
-            child: Text("View All", style: AppFont.medium_white_12.copyWith(color: AppColors.red)),
-          )
+            child: Text(
+              "View All",
+              style: AppFont.medium_white_12.copyWith(color: AppColors.red),
+            ),
+          ),
         ],
       ),
     );
@@ -100,7 +104,9 @@ class _WidgetHomePostersState extends State<WidgetHomePosters> {
                   begin: Alignment.centerLeft,
                   end: Alignment.centerRight,
                   colors: [
-                    Colors.black.withValues(alpha: 0.80), // Dark fade on the left
+                    Colors.black.withValues(
+                      alpha: 0.80,
+                    ), // Dark fade on the left
                     Colors.transparent,
                   ],
                 ),
@@ -121,7 +127,9 @@ class _WidgetHomePostersState extends State<WidgetHomePosters> {
                   begin: Alignment.centerRight,
                   end: Alignment.centerLeft,
                   colors: [
-                    Colors.black.withValues(alpha: 0.80), // Dark fade on the right
+                    Colors.black.withValues(
+                      alpha: 0.80,
+                    ), // Dark fade on the right
                     Colors.transparent,
                   ],
                 ),
@@ -160,8 +168,12 @@ class WidgetItemPoster extends StatelessWidget {
               ),
             ),
             WidgetSpacer(height: 4),
-            Text(item.title,
-                style: AppFont.regular_white_12, maxLines: 2, overflow: TextOverflow.ellipsis),
+            Text(
+              item.title,
+              style: AppFont.regular_white_12,
+              maxLines: 2,
+              overflow: TextOverflow.ellipsis,
+            ),
             WidgetSpacer(height: 2),
             Text(item.subTitle, style: AppFont.regular_white_10),
           ],
@@ -170,8 +182,8 @@ class WidgetItemPoster extends StatelessWidget {
     );
   }
 
-  void openMovieDetails(MovieEntity movie, BuildContext context) {
-    // TODO:
+  void openMovieDetails(MovieDetailEntity movie, BuildContext context) {
+    Navigator.pushNamed(context, AppRouter.MOVIE, arguments: movie);
   }
 }
 
@@ -179,13 +191,13 @@ class ItemPosterVM {
   late String title;
   late String subTitle;
   late String photo;
-  late MovieEntity movie;
+  late MovieDetailEntity movie;
 
   ItemPosterVM(this.photo, this.title, this.subTitle);
 
   ItemPosterVM.fromMovie(this.movie) {
-    title = movie.name;
-    subTitle = movie.tags.join(" - ");
-    photo = movie.thumb;
+    title = movie.detail.name;
+    subTitle = movie.detail.tags.join(" - ");
+    photo = movie.detail.thumb;
   }
 }
