@@ -19,7 +19,8 @@ part "authentication_state.dart";
 //
 // bloc to check if user is authenticated or not
 //
-class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> {
+class AuthenticationBloc
+    extends Bloc<AuthenticationEvent, AuthenticationState> {
   final FirebaseAuth _firebaseAuth = FirebaseAuth.instance;
   final GoogleSignIn _googleSignIn = GoogleSignIn();
 
@@ -32,7 +33,10 @@ class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> 
   AuthenticationState get initialState => Uninitialized();
 
   /// Handles the `AppStarted` event, checking if the user is signed in.
-  Future<void> _mapAppStartedToState(AppStarted event, Emitter<AuthenticationState> emit) async {
+  Future<void> _mapAppStartedToState(
+    AppStarted event,
+    Emitter<AuthenticationState> emit,
+  ) async {
     try {
       bool isSignedIn = await sl<IsSignedInUsecase>().call();
 
@@ -55,18 +59,21 @@ class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> 
   }
 
   /// Handles the `LoggedIn` event, fetching and setting the authenticated user state.
-  Future<void> _mapLoggedInToState(LoggedIn event, Emitter<AuthenticationState> emit) async {
+  Future<void> _mapLoggedInToState(
+    LoggedIn event,
+    Emitter<AuthenticationState> emit,
+  ) async {
     final user = await sl<GetUserUseCase>().call();
     emit(Authenticated(user.toString()));
   }
 
   /// Handles the `LoggedOut` event, signing out from Firebase and Google.
-  Future<void> _mapLoggedOutToState(LoggedOut event, Emitter<AuthenticationState> emit) async {
+  Future<void> _mapLoggedOutToState(
+    LoggedOut event,
+    Emitter<AuthenticationState> emit,
+  ) async {
     emit(Unauthenticated());
     // Sign out from both FirebaseAuth and GoogleSignIn asynchronously.
-    Future.wait([
-      _firebaseAuth.signOut(),
-      _googleSignIn.signOut(),
-    ]);
+    Future.wait([_firebaseAuth.signOut(), _googleSignIn.signOut()]);
   }
 }
