@@ -4,7 +4,10 @@
  * @ Message: 🎯 Happy coding and Have a nice day! 🌤️
  */
 
-import 'package:cinema_booking/data/models/responce/home_responce.dart';
+import 'package:cinema_booking/data/models/response/all_movie_by_type_response.dart';
+import 'package:cinema_booking/data/models/response/booking_time_slot_by_cinema_response.dart';
+import 'package:cinema_booking/data/models/response/home_response.dart';
+import 'package:cinema_booking/data/models/response/list_seat_slot_by_seat_type_response.dart';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
 
@@ -16,4 +19,13 @@ abstract class RestClient {
 
   @GET('/homeData')
   Future<HomeModelResponse> getHomeData();
+
+  @GET("/bookingTime")
+  Future<List<BookingTimeSlotByCinemaResponse>> getBookingTimeSlotByCine();
+
+  @GET("/bookSeat")
+  Future<ListSeatModelResponse> getListSeatSlotBySeatType();
+
+  @GET("/allMovie")
+  Future<AllMoviesModelResponse> getAllMoviesByType();
 }
