@@ -8,27 +8,39 @@ import 'package:flutter/material.dart';
 
 class AppColors {
   // primary and background colors
-  static const primary = Color(0xffF9C931);
-  static const defaultColor = Color(0xFFC012A0);
-  static const lightBackground = Color(0xffF2F2F2);
-  static const darkBackground = Color(0xff0D1118);
-  static const white = Color(0xffF6F6F6);
-  static const black = Color(0xff000000);
-  static const black2 = Color(0xFF333333);
-  static const transparent = Color(0x00000000);
-  static const grey = Color(0xffBEBEBE);
-  static const greyLight = Color(0xff999999);
-  static const gray3 = Color(0xFFF4F4F4);
-  static const gray4 = Color(0xFF666666);
-  static const gray5 = Color(0xFFC1C1C1);
-  static const gray6 = Color(0xFF707070);
-  static const gray7 = Color(0xFFDDDDDD);
-  static const greyDark = Color(0xff666666);
-  static const red = Color(0xffFF1D25);
-  static const blue = Color(0xff4267B2);
-  static const pink = Color(0xFFC012A0);
+  static const Color primary = Color(0xffF9C931);
+  static const Color defaultColor = Color(0xFFC012A0);
+  static const Color lightBackground = Color(0xffF2F2F2);
+  static const Color darkBackground = Color(0xff0D1118);
+  static const Color white = Color(0xffF6F6F6);
+  static const Color white1 = Color(0xFFFFE5EF);
+  static const Color black = Color(0xff000000);
+  static const Color black2 = Color(0xFF333333);
+  static const Color transparent = Color(0x00000000);
+  static const Color grey = Color(0xffBEBEBE);
+  static const Color gray1 = Color(0xFF999999);
+  static const Color greyLight = Color(0xff999999);
+  static const Color gray3 = Color(0xFFF4F4F4);
+  static const Color gray4 = Color(0xFF666666);
+  static const Color gray5 = Color(0xFFC1C1C1);
+  static const Color gray6 = Color(0xFF707070);
+  static const Color gray7 = Color(0xFFDDDDDD);
+  static const Color greyDark = Color(0xff666666);
+  static const Color orange = Color(0xFFEB8D2F);
+  static const Color red = Color(0xffFF1D25);
+  static const Color red2 = Color(0xFFD05045);
+  static const Color blue = Color(0xff4267B2);
+  static const Color pink = Color(0xFFC012A0);
+  static const Color green = Color(0xFF2EC492);
+  static const Color green2 = Color(0xFF8CC153);
+  static const Color gift1 = Color(0xFFFCCAC6);
+  static const Color gift2 = Color(0x2687D57E);
+  static const Color seatSlotBgBooked = Color(0xFFEBEBEB);
+  static const Color seatSlotBg = Color(0xFFFFFFFF);
+  static const Color timeSlotBg = Color(0xFFFCFCFC);
+  static const Color timeSlotBorder = Color(0xFFE5E5E5);
 
-  static const textLight = Color(0xffC6C6C6);
+  static const Color textLight = Color(0xffC6C6C6);
 
   // gradient splash
   static const List<Color> splashColor = [
