@@ -91,7 +91,9 @@ class AppFont {
     color: AppColors.black,
   );
 
-  static final regular_default = regular.copyWith(color: AppColors.defaultColor);
+  static final regular_default = regular.copyWith(
+    color: AppColors.defaultColor,
+  );
   static final regular_default_10 = regular_default.copyWith(fontSize: 10);
   static final regular_default_12 = regular_default.copyWith(fontSize: 12);
 
@@ -105,7 +107,9 @@ class AppFont {
   static final regular_gray1_10 = regular_gray1.copyWith(fontSize: 10);
   static final regular_gray1_12 = regular_gray1.copyWith(fontSize: 12);
 
-  static final regular_gray1_50 = regular.copyWith(color: AppColors.greyLight.withAlpha(50));
+  static final regular_gray1_50 = regular.copyWith(
+    color: AppColors.greyLight.withAlpha(50),
+  );
   static final regular_gray1_50_9 = regular_gray1_50.copyWith(fontSize: 9);
   static final regular_gray1_50_12 = regular_gray1_50.copyWith(fontSize: 12);
 
@@ -120,7 +124,9 @@ class AppFont {
   static final regular_gray4_12 = regular_gray4.copyWith(fontSize: 12);
   static final regular_gray4_14 = regular_gray4.copyWith(fontSize: 14);
 
-  static final regular_gray4_40 = regular.copyWith(color: AppColors.gray4.withAlpha(40));
+  static final regular_gray4_40 = regular.copyWith(
+    color: AppColors.gray4.withAlpha(40),
+  );
   static final regular_gray4_40_12 = regular_gray4_40.copyWith(fontSize: 12);
 
   static final regular_gray5 = regular.copyWith(color: AppColors.gray5);
@@ -147,7 +153,9 @@ class AppFont {
   static final medium_default_14 = medium_default.copyWith(fontSize: 14);
   static final medium_default_16 = medium_default.copyWith(fontSize: 16);
 
-  static final medium_white_70 = medium.copyWith(color: AppColors.white.withAlpha(70));
+  static final medium_white_70 = medium.copyWith(
+    color: AppColors.white.withAlpha(70),
+  );
   static final medium_white_70_14 = medium_white_70.copyWith(fontSize: 14);
 
   static final medium_gray4 = medium.copyWith(color: AppColors.gray4);
@@ -160,4 +168,23 @@ class AppFont {
   static final medium_blue = medium.copyWith(color: AppColors.blue);
   static final medium_blue_14 = medium_blue.copyWith(fontSize: 14);
   static final medium_blue_16 = medium_blue.copyWith(fontSize: 16);
+
+  // semibold
+  static final semibold_white = semibold.copyWith(color: AppColors.white);
+  static final semibold_white_10 = semibold_white.copyWith(fontSize: 10);
+  static final semibold_white_16 = semibold_white.copyWith(fontSize: 16);
+  static final semibold_white_18 = semibold_white.copyWith(fontSize: 18);
+
+  // oswald
+  static final oswald_regular = TextStyle(
+    fontFamily: 'Oswald',
+    fontWeight: FontWeight.w400,
+  );
+
+  static final oswald_regular_red2 = oswald_regular.copyWith(
+    color: AppColors.red2,
+  );
+  static final oswald_regular_red2_12 = oswald_regular_red2.copyWith(
+    fontSize: 12,
+  );
 }
