@@ -11,5 +11,6 @@ class AppUrls {
       'https://firebasestorage.googleapis.com/v0/b/spotify-flutter-4aa82.appspot.com/o/songs%2F';
   static const temp = 'Son-Tung-';
   static const mediaAlt = 'alt=media';
-  static const defaultAvatar = 'https://cdn-icons-png.flaticon.com/512/10542/10542486.png';
+  static const defaultAvatar =
+      'https://cdn-icons-png.flaticon.com/512/10542/10542486.png';
 }
