@@ -8,8 +8,5 @@ class SigninUserReq {
   final String email;
   final String password;
 
-  SigninUserReq({
-    required this.email,
-    required this.password,
-  });
+  SigninUserReq({required this.email, required this.password});
 }
