@@ -6,13 +6,8 @@ part of 'banner.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-BannerModel _$BannerModelFromJson(Map<String, dynamic> json) => BannerModel(
-      id: json['id'] as String?,
-      url: json['image'] as String?,
-    );
+BannerModel _$BannerModelFromJson(Map<String, dynamic> json) =>
+    BannerModel(id: json['id'] as String?, url: json['image'] as String?);
 
 Map<String, dynamic> _$BannerModelToJson(BannerModel instance) =>
-    <String, dynamic>{
-      'id': instance.id,
-      'image': instance.url,
-    };
+    <String, dynamic>{'id': instance.id, 'image': instance.url};
