@@ -6,13 +6,13 @@
 
 import 'package:cinema_booking/core/configs/assets/app_vectors.dart';
 import 'package:cinema_booking/domain/entities/genres/genres.dart';
-import 'package:cinema_booking/domain/entities/movies/movies.dart';
+import 'package:cinema_booking/domain/entities/response/home.dart';
 import 'package:cinema_booking/presentation/home/widgets/widget_home_poster.dart';
 import 'package:flutter/material.dart';
 
 class WidgetHomeEvents extends StatelessWidget {
   final GenresEntity genres;
-  final List<MovieEntity> movies;
+  final List<MovieDetailEntity> movies;
 
   const WidgetHomeEvents({
     super.key,
