@@ -4,6 +4,7 @@ import 'package:cinema_booking/common/widgets/space/widget_spacer.dart';
 import 'package:cinema_booking/core/configs/assets/app_images.dart';
 import 'package:cinema_booking/core/configs/theme/app_color.dart';
 import 'package:cinema_booking/core/configs/theme/app_font.dart';
+import 'package:cinema_booking/presentation/router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -38,23 +39,23 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
   }
 
   void _setupAnimations() {
-    _avatarController = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 2),
-    )..repeat(reverse: true);
+    _avatarController = AnimationController(vsync: this, duration: const Duration(seconds: 2))
+      ..repeat(reverse: true);
 
-    _avatarGlow = Tween<double>(begin: 1.0, end: 1.2).animate(
-      CurvedAnimation(parent: _avatarController, curve: Curves.easeInOut),
-    );
+    _avatarGlow = Tween<double>(
+      begin: 1.0,
+      end: 1.2,
+    ).animate(CurvedAnimation(parent: _avatarController, curve: Curves.easeInOut));
 
     _notificationController = AnimationController(
       vsync: this,
       duration: const Duration(milliseconds: 400),
     );
 
-    _notificationShake = Tween<double>(begin: 0.0, end: 4.0).animate(
-      CurvedAnimation(parent: _notificationController, curve: Curves.elasticIn),
-    );
+    _notificationShake = Tween<double>(
+      begin: 0.0,
+      end: 4.0,
+    ).animate(CurvedAnimation(parent: _notificationController, curve: Curves.elasticIn));
   }
 
   void _triggerNotificationShake() {
@@ -84,24 +85,13 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
       padding: const EdgeInsets.symmetric(horizontal: 16),
       color: AppColors.darkBackground,
       height: 70,
-      child: Row(
-        children: <Widget>[
-          _buildAvatarAndName(),
-          _buildNotificationIcon(),
-        ],
-      ),
+      child: Row(children: <Widget>[_buildAvatarAndName(), _buildNotificationIcon()]),
     );
   }
 
   Widget _buildAvatarAndName() {
     return Expanded(
-      child: Row(
-        children: [
-          _buildAnimatedAvatar(),
-          WidgetSpacer(width: 12),
-          _buildAnimatedText(),
-        ],
-      ),
+      child: Row(children: [_buildAnimatedAvatar(), WidgetSpacer(width: 12), _buildAnimatedText()]),
     );
   }
 
@@ -118,12 +108,9 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
               color: AppColors.primary.withValues(alpha: 0.6),
               blurRadius: 10,
               spreadRadius: 2,
-            )
+            ),
           ],
-          border: Border.all(
-            color: AppColors.white,
-            width: 1.0,
-          ),
+          border: Border.all(color: AppColors.white, width: 1.0),
           image: DecorationImage(image: AssetImage(AppImages.avatar)),
         ),
       ),
@@ -155,13 +142,11 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisAlignment: MainAxisAlignment.center,
       children: <Widget>[
-        Text(
-          'Chung Nguyen Thanh',
-          style: AppFont.medium_white_16.copyWith(fontSize: 15),
-        ),
+        Text('Chung Nguyen Thanh', style: AppFont.medium_white_16.copyWith(fontSize: 15)),
         GestureDetector(
           onTap: () {
-            _clickSignOut();
+            Navigator.of(context).pushNamed(AppRouter.LIST_TICKETS);
+            // _clickSignOut();
           },
           child: Opacity(
             opacity: 0.5,
@@ -182,13 +167,7 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
       key: ValueKey(text),
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Text(
-          text,
-          style: AppFont.medium_white_16,
-          textAlign: TextAlign.left,
-        ),
-      ],
+      children: <Widget>[Text(text, style: AppFont.medium_white_16, textAlign: TextAlign.left)],
     );
   }
 
@@ -196,10 +175,7 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
     return AnimatedBuilder(
       animation: _notificationController,
       builder: (context, child) {
-        return Transform.translate(
-          offset: Offset(_notificationShake.value, 0.0),
-          child: child,
-        );
+        return Transform.translate(offset: Offset(_notificationShake.value, 0.0), child: child);
       },
       child: GestureDetector(
         onTap: () {
@@ -214,14 +190,10 @@ class _WidgetHomeToolbarState extends State<WidgetHomeToolbar> with TickerProvid
                 color: AppColors.red.withValues(alpha: 0.50),
                 blurRadius: 8,
                 spreadRadius: 2,
-              )
+              ),
             ],
           ),
-          child: const Icon(
-            Icons.notifications,
-            color: AppColors.white,
-            size: 24,
-          ),
+          child: const Icon(Icons.notifications, color: AppColors.white, size: 24),
         ),
       ),
     );
